fixing crashes after OSErrors. This commit only avoids the crash of the script. It does not fix the underlying problem, that cgroup-directory is missing.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@12383 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ scripts/benchmark/cgroups.py
@@ -170,6 +170,9 @@ def killAllTasksInCgroup(cgroup):
 
 def removeCgroup(cgroup):
     if cgroup:
+        if not os.path.exists(cgroup):
+            logging.warning('Cannot remove CGroup {0}, because it does not exist.'.format(cgroup))
+            return
         assert os.path.getsize(os.path.join(cgroup, 'tasks')) == 0
         try:
             os.rmdir(cgroup)
